chore: separate member modules by layers
