+++ module/domain/member/event-listener/build.gradle.kts
@@ -1,0 +1,16 @@
+plugins {
+    alias(libs.plugins.spring.boot)
+    alias(libs.plugins.spring.dependency.management)
+}
+
+dependencies {
+//    implementation(project(":core-security"))
+//
+//    implementation(project(":common-mysql"))
+//    implementation(project(":common-redis"))
+//
+//    implementation("org.springframework.boot:spring-boot-starter-data-jpa")
+
+    testImplementation("org.springframework.boot:spring-boot-starter-test")
+    testImplementation("org.jetbrains.kotlin:kotlin-test-junit5")
+}
