+++ module/domain/member/service/src/test/kotlin/com/ecommerce/MemberUnitTest.kt
@@ -1,6 +1,5 @@
 package com.ecommerce
 
-import com.ecommerce.service.Member
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
